refactor: move build logic to library
